requirements.txt
Django>=4.2,<5
gunicorn>=21.0
whitenoise>=6.0
python-telegram-bot>=20.0
beautifulsoup4>=4.13.4
httpx>=0.27.0
loguru>=0.7.0
pydantic>=2.9.2
requests>=2.32.3
apscheduler>=3.10.0
tomli>=2.0.0
